backend/requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
newsapi-python
transformers
torch
xgboost
scikit-learn
joblib
python-dateutil